\name{IntegerSort-package}
\alias{IntegerSort-package}
\alias{IntegerSort}
\docType{package}

\title{Sorting Methods For Integer Vector}
\description{
This package uses RAM based algorithm such as radix sort and counting sort and is designed to manipulate large data. For opposite target, insertion sort algorithm is added to operate small data set. Such algorithms give more performance when it used with uniform data and are considered as low level tools (C/C++ resources ) which leads more vigilance and good practice of user.            } 
\author{Med Ezzeddine MACHERKI}
